Renamed world to another_crab and fixed victory condition
Renamed world to another_crab to make it easier to identify.

Made fixes to get location/region rules working.

Made fixes to get victory condition working.
