+++ worlds/another_crab/names/location_names.py
